requirements.txt
azure-storage-blob
openai
azure-functions
pandas
requests